Implemented POSIX support for software clock
A new option has been added, -d. This will allow the user to pass a
path to ptp4l to a POSIX clock. If software mode is enabled, and the
path is set, ptp4l will try to open the path as a POSIX clock. This
allows for custom clocks to implement the posix clock interface and
still be able to utilize ptp time synchronization.

Signed-off-by: Dimitrios Katsaros <patcherwork@gmail.com>